Changes in Exception, fixed some remarks

+++ Service/src/main/java/com/epam/esm/service/impl/GiftCertificateServiceImpl.java
@@ -1,11 +1,13 @@
 package com.epam.esm.service.impl;
 
 import com.epam.esm.dao.GiftCertificateDao;
-import com.epam.esm.dao.TagDao;
 import com.epam.esm.entity.GiftCertificate;
 import com.epam.esm.entity.Tag;
+import com.epam.esm.exception.EntityException;
 import com.epam.esm.service.GiftCertificateService;
+import com.epam.esm.service.TagService;
 import com.epam.esm.validator.GiftCertificateValidator;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,30 +15,42 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
+import static com.epam.esm.exception.ExceptionCode.*;
+
+@Log4j2
 @Service
 public class GiftCertificateServiceImpl implements GiftCertificateService {
     private final GiftCertificateDao giftCertificateDao;
-    private final TagDao tagDao;
+    private final TagService tagService;
     private final GiftCertificateValidator giftCertificateValidator;
-    private final static String FIELD_NAME = "name";
-    private final static String FIELD_DESCRIPTION = "description";
-    private final static String FIELD_PRICE = "price";
-    private final static String FIELD_DURATION = "duration";
+    private static final String FIELD_NAME = "name";
+    private static final String FIELD_DESCRIPTION = "description";
+    private static final String FIELD_PRICE = "price";
+    private static final String FIELD_DURATION = "duration";
 
     @Autowired
-    public GiftCertificateServiceImpl(GiftCertificateDao giftCertificateDao, TagDao tagDao,
+    public GiftCertificateServiceImpl(GiftCertificateDao giftCertificateDao, TagService tagService,
                                       GiftCertificateValidator giftCertificateValidator) {
         this.giftCertificateDao = giftCertificateDao;
-        this.tagDao = tagDao;
         this.giftCertificateValidator = giftCertificateValidator;
+        this.tagService = tagService;
     }
 
     @Override
-    public void create(GiftCertificate giftCertificate) {
-        if (giftCertificateValidator.isValid(giftCertificate)){
-        giftCertificateDao.create(giftCertificate);
+    @Transactional
+    public GiftCertificate create(GiftCertificate giftCertificate) {
+        if (!giftCertificateValidator.isValid(giftCertificate)){
+            throw new EntityException(NOT_VALID_GIFT_CERTIFICATE_DATA.getErrorCode());
         }
+        Long newId = giftCertificateDao.create(giftCertificate);
+        Set<Tag> tags = giftCertificate.getTags();
+        if(giftCertificateValidator.isTagsAttachedToCertificate(tags)){
+            tags.stream().allMatch(tagService::isTagValid);
+            addTagsToCertificate(tags, newId);
+        }
+        return findById(newId);
     }
 
     @Override
@@ -47,55 +61,74 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     @Override
     public GiftCertificate findById(Long id) {
         GiftCertificate giftCertificate = giftCertificateDao.findById(id);
-        if (giftCertificate != null){
-            giftCertificate.setTags(tagDao.findAllTagInCertificate(giftCertificate.getId()));
+        if (giftCertificate == null){
+            throw new EntityException(GIFT_CERTIFICATE_NOT_FOUND.getErrorCode());
         }
         return giftCertificate;
     }
 
     @Override
-    @Transactional
     public void delete(long id) {
-        giftCertificateDao.delete(id);
-        tagDao.deleteAllTagFromCertificate(id);
+         if (isGiftCertificateExist(id)){
+             giftCertificateDao.delete(id);
+         }
     }
 
     @Override
-    public List<GiftCertificate> findAllCertificateByTag(String tagName){
-        return giftCertificateDao.findAllCertificateByTag(tagName);
+    @Transactional
+    public GiftCertificate addTagToCertificate(Tag tag, long idCertificate){
+        if(isGiftCertificateExist(idCertificate) && isTagReadyToCreate(tag)) {
+            tagService.create(tag);
+        }
+        tagService.addTagToCertificate(tagService.findTagByName(tag.getName()), idCertificate);
+        giftCertificateDao.updateDate(idCertificate);
+        return giftCertificateDao.findById(idCertificate);
     }
 
     @Override
-    public void addTagToCertificate(Tag tag, long idCertificate){
-        tagDao.addTagToCertificate(tag, idCertificate);
+    @Transactional
+    public void addTagsToCertificate(Set<Tag> tags, long idCertificate){
+        tagService.createTags(tags);
+        List<Tag> newTags = tagService.findTagsByName(tags);
+        tagService.addTagSToCertificate(newTags, idCertificate);
     }
 
     @Override
-    public void deleteTagFromCertificate(Tag tag, long idCertificate){
-        tagDao.deleteTagFromCertificate(tag, idCertificate);
-    }
-
-    @Override
-    public List<GiftCertificate> findCertificateByName(String name){
-        return giftCertificateDao.findByName(name);
-    }
-
-    @Override
-    public List<GiftCertificate> findCertificateByDescription(String description){
-        return giftCertificateDao.findByName(description);
+    @Transactional
+    public GiftCertificate deleteTagFromCertificate(Tag tag, long idCertificate){
+        if(isTagCanBeDeletedFromCertificate(tag, idCertificate)){
+            tagService.deleteTagFromCertificate(tagService.findTagByName(tag.getName()), idCertificate);
+        }
+        giftCertificateDao.updateDate(idCertificate);
+        return giftCertificateDao.findById(idCertificate);
     }
 
     @Override
     public GiftCertificate update(Long id, GiftCertificate updatedGiftCertificate) {
-        GiftCertificate currentGiftCertificate = giftCertificateDao.findById(id);
-        Map<String,Object> updatedFields = getUpdatedField(updatedGiftCertificate, currentGiftCertificate);
-        giftCertificateDao.update(updatedGiftCertificate, updatedFields);
-        return null;
+        if(isGiftCertificateValid(updatedGiftCertificate) && isGiftCertificateExist(id)){
+            GiftCertificate currentGiftCertificate = giftCertificateDao.findById(id);
+            Map<String,Object> updatedFields = getUpdatedField(updatedGiftCertificate, currentGiftCertificate);
+            giftCertificateDao.update(id, updatedFields);
+        }
+        return giftCertificateDao.findById(id);
+    }
+
+    @Override
+    public List<GiftCertificate> findByAttributes(String tagName, String searchPart, String sortingField,
+                                                  String orderSort, String search) {
+        if (search ==  null){
+            return giftCertificateDao.findAll();
+        }
+        if (giftCertificateValidator.isGiftCertificateFieldValid(sortingField)
+                && giftCertificateValidator.isOrderSortValid(orderSort)) {
+            return giftCertificateDao.findByAttributes(tagName, searchPart, sortingField, orderSort);
+        }
+        throw new EntityException(WRONG_FIND_PARAMETERS.getErrorCode());
     }
 
     private Map<String,Object> getUpdatedField(GiftCertificate updatedGiftCertificate,
                                                GiftCertificate currentGiftCertificate){
-        Map<String,Object> updatedFields = new HashMap<String,Object>();
+        Map<String,Object> updatedFields = new HashMap<>();
         if(updatedGiftCertificate.getName() !=null &&
                 !updatedGiftCertificate.getName().equals(currentGiftCertificate.getName())){
             updatedFields.put(FIELD_NAME, updatedGiftCertificate.getName());
@@ -112,9 +145,28 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
                 !updatedGiftCertificate.getPrice().equals(currentGiftCertificate.getPrice())){
             updatedFields.put(FIELD_PRICE, updatedGiftCertificate.getPrice());
         }
-
         return updatedFields;
     }
 
+    private boolean isGiftCertificateValid(GiftCertificate giftCertificate){
+        if(!giftCertificateValidator.isValid(giftCertificate)){
+            throw new EntityException(NOT_VALID_GIFT_CERTIFICATE_DATA.getErrorCode());
+        }
+        return true;
+    }
 
+    private boolean isGiftCertificateExist(Long id){
+        if(giftCertificateDao.findById(id) == null){
+            throw new EntityException(GIFT_CERTIFICATE_NOT_FOUND.getErrorCode());
+        }
+        return true;
+    }
+
+    private boolean isTagReadyToCreate(Tag tag){
+        return tagService.isTagValid(tag) && tagService.findTagByName(tag.getName()) == null;
+    }
+
+    private boolean isTagCanBeDeletedFromCertificate(Tag tag , long idCertificate){
+        return tagService.isTagValid(tag) && isGiftCertificateExist(idCertificate) && tagService.isTagExist(tag);
+    }
 }
