+++ Service/src/main/java/com/epam/esm/service/impl/TagServiceImpl.java
@@ -2,14 +2,17 @@ package com.epam.esm.service.impl;
 
 import com.epam.esm.dao.TagDao;
 import com.epam.esm.entity.Tag;
-import com.epam.esm.exception.EntityAlreadyExistsException;
-import com.epam.esm.exception.EntityNotFoundException;
+import com.epam.esm.exception.EntityException;
 import com.epam.esm.service.TagService;
 import com.epam.esm.validator.TagValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Set;
+
+import static com.epam.esm.exception.ExceptionCode.NOT_VALID_TAG_DATA;
+import static com.epam.esm.exception.ExceptionCode.TAG_NOT_FOUND;
 
 @Service
 public class TagServiceImpl implements TagService {
@@ -23,21 +26,21 @@ public class TagServiceImpl implements TagService {
     }
 
     @Override
-    public void create(Tag tag) {
-        if(tagDao.findTagByName(tag.getName()) != null){
-            throw new EntityAlreadyExistsException("message.tagAlreadyExists");
+    public Tag create(Tag tag) {
+        if(isTagValid(tag) && tagDao.findTagByName(tag.getName()) == null){
+            Long idCreatedTag = tagDao.create(tag);
+            return findTag(idCreatedTag);
         }
-        if(tagValidator.isValid(tag)){
-            tagDao.create(tag);
-        }
+        return findTagByName(tag.getName());
     }
 
     @Override
     public Tag findTag(Long id) {
-        if (tagDao.findTag(id) == null){
-            throw new EntityNotFoundException("message.tagNotFound");
+        Tag tag = tagDao.findTag(id);
+        if (tag == null){
+            throw new EntityException(TAG_NOT_FOUND.getErrorCode());
         }
-        return tagDao.findTag(id);
+        return tag;
     }
 
     @Override
@@ -47,7 +50,55 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public void delete(Long id) {
-        tagDao.delete(id);
+        if (tagDao.findTag(id) != null){
+            tagDao.delete(id);
+        }
+        else throw new EntityException(TAG_NOT_FOUND.getErrorCode());
     }
 
+    @Override
+    public boolean isTagExist(Tag tag){
+        if(tagDao.findTagByName(tag.getName()) == null){
+            throw new EntityException(TAG_NOT_FOUND.getErrorCode());
+        }
+        return true;
+    }
+
+    @Override
+    public void addTagToCertificate(Tag tag, Long idCertificate) {
+        tagDao.addTagToCertificate(tag,idCertificate);
+    }
+
+    @Override
+    public Tag findTagByName(String name) {
+        return tagDao.findTagByName(name);
+    }
+
+    @Override
+    public void deleteTagFromCertificate(Tag tag, Long idCertificate) {
+        tagDao.deleteTagFromCertificate(tag, idCertificate);
+    }
+
+    @Override
+    public List<Tag> findTagsByName(Set<Tag> tags) {
+        return tagDao.findTagsByName(tags);
+    }
+
+    @Override
+    public void createTags(Set<Tag> tags) {
+        tagDao.createTags(tags);
+    }
+
+    @Override
+    public void addTagSToCertificate(List<Tag> tags, Long idCertificate) {
+        tagDao.addTagSToCertificate(tags, idCertificate);
+    }
+
+    @Override
+    public boolean isTagValid(Tag tag){
+        if(!tagValidator.isValid(tag)){
+            throw new EntityException(NOT_VALID_TAG_DATA.getErrorCode());
+        }
+        return true;
+    }
 }
